Add fixes for ESP32-based azure_ca_recovery sample to build with ESP-IDF 5 and enable it in gate checks

+++ demos/common/azure_ca_recovery/azure_ca_recovery_mbedtls_rsa_verify.c
@@ -4,6 +4,7 @@
 #include "azure_ca_recovery_rsa_verify.h"
 
 #include "azure_iot_config.h"
+#include "azure_iot_jws.h"
 #include "azure/core/az_base64.h"
 
 #include "mbedtls/base64.h"
@@ -69,7 +70,7 @@ AzureIoTResult_t AzureIoTSample_RS256Verify( uint8_t * pucInput,
 
     if( az_result_failed( xCoreResult ) )
     {
-        AZLogError( ( "[RSA] Base64 decode failed: 0x%08x", xCoreResult ) );
+        AZLogError( ( "[RSA] Base64 decode failed: 0x%08x", ( unsigned int ) xCoreResult ) );
         return eAzureIoTErrorFailed;
     }
